Add files via upload
This repository contains the source code for the official website of Cattleya's Cakes, a virtual bakery offering a delectable array of cakes and desserts. Explore our mouthwatering product showcase, indulge in our delightful blog posts, and discover what our satisfied clients have to say in our review section. With a responsive design and captivating visuals, our website promises to sweeten your online experience. Join us as we celebrate the joy of baking and the art of dessert creation! 🎂🍰"

Feel free to tailor this description to better suit your style and the specific features of your website!